added main, cmake lib to quarry, new env
