+++ .env.example
@@ -1,0 +1,1 @@
+MASSIVE_API_KEY = massive-key
