--- quarry/.env.example
@@ -1,1 +1,0 @@
-POLYGON_API_KEY = massive-key
